Refactors ShallowWaterModels module and cleans up test
